Fetch raw list from Pocket
So far, the app simply connects to Pocket and downloads a batch of
saved links.

* Filter links that should be persisted
* Take the original URL of those links
* Curl the contents and store (links -dump $url > file.txt)
* Remember the last one persisted to allow regular runs

Setup
=====

    export CONSUMER_KEY=<key>
    export ACCESS_TOKEN=<token>

The Consumer Key is obtained by creating a new application in the
GetPocket web interface at https://getpocket.com/developer/apps/new

Now run

    ./retrieve_token $CONSUMER_KEY

And follow the instructions.  You'll need to open a link in your browser
(where you are assumed to be logged into Pocket) and authorize the
application.  The script will then print the access_token that you
should set in the environment variable above.

Signed-off-by: Galo Navarro <anglorvaroa@gmail.com>